requirements.txt
python-chess
langchain
langchain-deepseek-official
langchain-google-genai
google-generativeai
fastapi
uvicorn